# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Translalia API
  version: 2026-03-12
  description: >
    Current-code-first catalog of the Next.js API routes in `translalia-web/src/app/api`.
    Stable/high-use routes have fuller request and response schemas; lower-signal internal
    routes are cataloged with generic object shapes so they remain discoverable.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/notebook/suggestions:
    post:
      tags: [notebook]
      summary: Run identify, adjust, or personalize notebook suggestions
      x-classification: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotebookSuggestionRequest"
      responses:
        "200":
          description: Notebook suggestion step result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotebookSuggestionResponse"
...
  /api/diary/completed-poems:
    get:
      tags: [diary]
      summary: Fetch completed poems archive
      x-classification: public
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - in: query
          name: beforeCreatedAt
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: beforeId
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Completed poem page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiaryCompletedPoemsResponse"
...
components:
...
  schemas:
...
    NotebookSuggestionRequest:
      type: object
      required: [threadId, step, sourcePoem, translationPoem]
      properties:
        threadId:
          type: string
          format: uuid
        step:
          type: string
          enum: [identify, adjust, personalize]
        sourcePoem:
          type: string
        translationPoem:
          type: string
        sourceLanguage:
          type: string
        targetLanguage:
          type: string
        translationDiary:
          type: string
        lineNotes:
          type: object
          additionalProperties:
            type: string
        formalFeatures:
          type: object
          additionalProperties: true
        selectedLines:
          type: array
          items:
            type: integer
            minimum: 0
    NotebookSuggestionResponse:
      type: object
      properties:
        ok:
          type: boolean
        step:
          type: string
          enum: [identify, adjust, personalize]
        formalFeatures:
          type: object
          additionalProperties: true
        adjustments:
          type: object
          additionalProperties: true
        personalized:
          type: object
          additionalProperties: true
        cached:
          type: boolean
        error:
          type: object
          additionalProperties: true
      required: [ok, step]
    DiaryCompletedPoem:
      type: object
      properties:
        thread_id:
          type: string
          format: uuid
        title:
          type: string
        thread_created_at:
          type: string
          format: date-time
        raw_poem:
          type: [string, "null"]
        workshop_lines:
          description: Saved workshop line array as returned by the RPC.
          oneOf:
            - type: "null"
            - type: array
              items:
                type: object
                additionalProperties: true
        notebook_notes:
          oneOf:
            - type: "null"
            - type: object
              additionalProperties: true
        expressYourView:
          type: [string, "null"]
          description: Student post-AI reflection from state.express_your_view.
        translationInsights:
          oneOf:
            - type: "null"
            - type: object
              description: Latest step-c snapshot from state.translation_insights.
              additionalProperties: true
        refineRhyme:
          oneOf:
            - type: "null"
            - type: object
              description: Refine & Rhyme artifacts from state.refine_rhyme.
              additionalProperties: true
        journey_summary_created_at:
          type: [string, "null"]
          format: date-time
        reflection_text:
          type: [string, "null"]
        insights:
          type: [array, "null"]
          items:
            type: string
        strengths:
          type: [array, "null"]
          items:
            type: string
        challenges:
          type: [array, "null"]
          items:
            type: string
        recommendations:
          type: [array, "null"]
          items:
            type: string
    DiaryCompletedPoemsResponse:
      type: object
      properties:
        ok:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/DiaryCompletedPoem"
        nextCursor:
          oneOf:
            - type: "null"
            - type: object
              properties:
                beforeCreatedAt:
                  type: string
                  format: date-time
                beforeId:
                  type: string
                  format: uuid
      required: [ok, items]
...